+++ wezterm/src/cli/set_tab_title.rs
@@ -1,4 +1,3 @@
-use crate::cli::resolve_pane_id;
 use clap::Parser;
 use mux::pane::PaneId;
 use mux::tab::TabId;
@@ -46,7 +45,7 @@ impl SetTabTitle {
             tab_id
         } else {
             // Find the current tab from the pane id
-            let pane_id = resolve_pane_id(&client, self.pane_id).await?;
+            let pane_id = client.resolve_pane_id(self.pane_id).await?;
             pane_id_to_tab_id
                 .get(&pane_id)
                 .copied()